fix: initialize some paddings and unions to 0 to get more deterministic parser output
This was added to more easily (binary) compare two parser outputs and detect if they are the same, affects the files in scripts/ and parser/utxoState.dat.
Maybe this commit should be reverted before release, as using memset() on anonymous unions is apparently undefined behavior (see https://stackoverflow.com/q/42361656).