feat: move script header data (txFirstSeen, txFirstSpent, typesSeen) to separate header files
This change allows to store the script header data for every fork separately. Performance should not be affected, unless script data like the pubkey is needed, which is usually not needed for analyses. One relevant implication of this change is that now addresses can be "unseen" on a chain, eg. an address appeared on BTC, but not on BCH.

BREAKING CHANGE: Reparse of existing parser output data is needed. Address.first_tx now returns Optional[blocksci.Tx] (before: blocksci.Tx)