fix: script header data is (sometimes) incorrect for forked chains in multi-chain mode
BREAKING CHANGE: Reparsing the blockchain is strongly recommended for existing multi-chain mode parsings.

This fixes a (rather uncommon) edge-case, where an address that existed before the fork is used with an equivalent type after the fork on the main chain.
E.g., a pubkey address pre-fork, and a multisig-pubkey address after-fork.
When parsing the forked chain, this address is correctly resolved as "existing", but this causes that ScriptHeader.txFirstSeen and ScriptHeader.typesSeen is not updated.